fix : server,server-action 함수 분리
